third_party/blink/web_tests/platform/generic/fast/forms/label/labels-contenteditable-expected.txt

ALERT: PASS: label passed the focus to the corresponding control
ALERT: PASS: focus should be sent if the label has contenteditable attribute
ALERT: PASS: keydown should be sent if the label has contenteditable attribute
ALERT: PASS: keypress should be sent if the label has contenteditable attribute
ALERT: PASS: keyup should be sent if the label has contenteditable attribute
ALERT: PASS: blur should be sent if the label has contenteditable attribute
ALERT: PASS: label passed the focus to the corresponding control
Test that labels with contenteditable attribute send focus and key events.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

text tAext
